requirements.txt
asgiref==3.5.2
bcrypt==4.0.0
boto3==1.24.87
botocore==1.27.87
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cloudinary==1.30.0
coreapi==2.3.3
coreschema==0.0.4
crispy-bootstrap5==0.7
cryptography==38.0.1
Django==3.2.15
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-environ==0.9.0
django-filter==22.1
django-relativedelta==1.1.2
django-storages==1.13.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonfield==3.1.0
jsons==1.6.3
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
paramiko==2.11.0
Pillow==9.2.0
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
typish==1.9.3
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
